error handling and responsiveness
