+++ app/dashboard/applications/data-table.tsx
@@ -119,7 +119,7 @@ export function DataTable<TData extends Application, TValue>({
                     ))}
                 </TableBody>
             </Table>
-            <div className="flex items-center justify-between space-x-4 py-4">
+            <div className="flex flex-col md:flex-row items-center justify-between space-x-4 py-4">
                 <div className="flex items-center space-x-1 px-4">
                     <span>
                         Page {table.getState().pagination.pageIndex + 1} of {table.getPageCount()}
